requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
cffi==1.14.4
chardet==3.0.4
cycler==0.10.0
discord.py==1.5.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
multidict==5.1.0
numpy==1.19.3
opencv-python==4.4.0.46
pandas== 1.1.5
Pillow==8.0.1
pip==20.3.3
pycparser==2.20
pygame==2.0.0
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
setuptools==49.2.1
six==1.15.0
typing-extensions==3.7.4.3
wheel==0.36.2
xlrd==1.2.0
yarl==1.6.3